Replace boost::lexical_cast with std::stoul or std::to_string.
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/branches/x64@6364 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ tdm_update/libtdm_update/Http/MirrorDownload.cpp
@@ -91,7 +91,7 @@ std::string MirrorDownload::GetCurrentMirrorName()
 
 void MirrorDownload::InitRandomizer()
 {
-    _rand.seed(std::chrono::system_clock::now().time_since_epoch().count());
+    _rand.seed(static_cast<unsigned long>(std::chrono::system_clock::now().time_since_epoch().count()));
 }
 
 } // namespace
